#47 Infra 모듈 생성 및 구조 수정
