+++ app/src/main/AndroidManifest.xml
@@ -36,29 +36,6 @@
                     android:scheme="kakao${NATIVE_APP_KEY}" />
             </intent-filter>
         </activity>
-
-        <receiver android:name=".alarm.AlarmReceiver"
-            android:enabled="true"
-            android:exported="false" >
-            <intent-filter>
-                <action android:name="com.same.alarm.ACTION_NAME" />
-            </intent-filter>
-        </receiver>
-
-        <receiver android:name="com.same.alarm.alarm.RescheduleAlarmReceiver"
-            android:enabled="true"
-            android:exported="false" >
-            <intent-filter>
-                <action android:name="android.intent.action.BOOT_COMPLETED" />
-                <action android:name="android.intent.action.LOCKED_BOOT_COMPLETED" />
-            </intent-filter>
-        </receiver>
-
-        <service
-            android:name="com.same.alarm.alarm.ForegroundAlarmService"
-            android:exported="false"
-            android:foregroundServiceType="dataSync" />
-
     </application>
 
 </manifest>
