+++ core/common/src/main/java/com/same/alarm/common/AlarmConfig.kt
@@ -2,7 +2,7 @@ package com.same.alarm.common
 
 import java.time.DayOfWeek
 
-object Constants {
+object AlarmConfig {
     val DAYS_OF_WEEK = DayOfWeek.entries.toTypedArray()
     val CATEGORIES = listOf("스터디", "기상", "미팅", "집안일")
 }
