+++ core/data/src/test/java/com/same/alarm/data/AlarmTest.kt
@@ -1,4 +1,4 @@
-package com.same.alarm
+package com.same.alarm.data
 
 import com.same.alarm.model.Alarm
 import org.junit.Test
@@ -10,7 +10,7 @@ import java.time.format.DateTimeFormatter
 class AlarmTest {
 
     @Test
-    fun testGetAlarmFirstTriggerMillis() {
+    fun 알람_첫_번째_트리거_시간_계산() {
         val alarm = Alarm(
             id = 1,
             time = LocalTime.of(9, 0),
