+++ feature/ring/build.gradle.kts
@@ -16,4 +16,5 @@ dependencies {
     implementation(project(":core:designsystem"))
     implementation(project(":core:domain"))
     implementation(project(":core:model"))
+    implementation(project(":core:common"))
 }
