+++ settings.gradle.kts
@@ -42,3 +42,4 @@ include(":feature:edit")
 include(":core:common")
 include(":feature:login")
 include(":core:datastore")
+include(":infra")
